+++ python/lsst/ts/salobj/topics/read_topic.py
@@ -240,7 +240,7 @@ class ReadTopic(BaseTopic):
         name,
         sal_prefix,
         max_history,
-        queue_len=100,
+        queue_len=DDS_READ_QUEUE_LEN,
         filter_ackcmd=True,
     ):
         super().__init__(salinfo=salinfo, name=name, sal_prefix=sal_prefix)
@@ -275,7 +275,18 @@ class ReadTopic(BaseTopic):
             if self.volatile
             else salinfo.domain.reader_qos
         )
-        self._reader = salinfo.subscriber.create_datareader(self._topic, qos)
+        # Command topics use a different a partition name than
+        # all other topics, including ackcmd, and the partition name
+        # is part of the publisher and subscriber.
+        # This split allows us to create just one subscriber and one publisher
+        # for each Controller or Remote:
+        # `Controller` only needs a cmd_subscriber and data_publisher,
+        # `Remote` only needs a cmd_publisher and data_subscriber.
+        if sal_prefix == "command_":
+            subscriber = salinfo.cmd_subscriber
+        else:
+            subscriber = salinfo.data_subscriber
+        self._reader = subscriber.create_datareader(self._topic, qos)
         # TODO DM-26411: replace ANY_INSTANCE_STATE with ALIVE_INSTANCE_STATE
         # once the OpenSplice issue 00020647 is fixed.
         read_mask = [
@@ -286,15 +297,9 @@ class ReadTopic(BaseTopic):
         if salinfo.index > 0:
             queries.append(f"{salinfo.name}ID = {salinfo.index}")
         if name == "ackcmd" and filter_ackcmd:
-            # TODO DM-25474: enable the identity test
-            # once all CSCs echo identity in their ackcmd topics
-            # (ts_salobj 6 and ts_sal 4.2 used everywhere).
-            # I tried to write a query that checks for
-            # identity='' or identity=salinfo.domain.origin
-            # but could not get OpenSplice to accept the empty string.
             queries += [
                 f"origin = {salinfo.domain.origin}",
-                # f"identity = '{salinfo.domain.identity}'",
+                f"identity = '{salinfo.identity}'",
             ]
         if queries:
             full_query = " AND ".join(queries)