+++ python/lsst/ts/salobj/topics/write_topic.py
@@ -96,7 +96,18 @@ class WriteTopic(BaseTopic):
             if self.volatile
             else salinfo.domain.writer_qos
         )
-        self._writer = salinfo.publisher.create_datawriter(self._topic, qos)
+        # Command topics use a different a partition name than
+        # all other topics, including ackcmd, and the partition name
+        # is part of the publisher and subscriber.
+        # This split allows us to create just one subscriber and one publisher
+        # for each Controller or Remote:
+        # `Controller` only needs a cmd_subscriber and data_publisher,
+        # `Remote` only needs a cmd_publisher and data_subscriber.
+        if sal_prefix == "command_":
+            publisher = salinfo.cmd_publisher
+        else:
+            publisher = salinfo.data_publisher
+        self._writer = publisher.create_datawriter(self._topic, qos)
         self._has_data = False
         self._data = self.DataType()
         self._has_priority = sal_prefix == "logevent_"
@@ -166,7 +177,7 @@ class WriteTopic(BaseTopic):
         self.data.private_sndStamp = base.current_tai()
         self.data.private_revCode = self.rev_code
         self.data.private_origin = self.salinfo.domain.origin
-        self.data.private_identity = self.salinfo.domain.identity
+        self.data.private_identity = self.salinfo.identity
         if self._seq_num_generator is not None:
             self.data.private_seqNum = next(self._seq_num_generator)
         # when index is 0 use the default of 0 and give senders a chance