+++ tests/test_salobj_to_either.py
@@ -47,7 +47,7 @@ index_gen = salobj.index_generator()
 
 class SALPYTestCase(asynctest.TestCase):
     def setUp(self):
-        salobj.set_random_lsst_dds_domain()
+        salobj.set_random_lsst_dds_partition_prefix()
         self.datadir = pathlib.Path(__file__).resolve().parent / "data"
         self.index = next(index_gen)
 
@@ -96,7 +96,7 @@ class SALPYTestCase(asynctest.TestCase):
                     ackcmd = await remote.cmd_setLogLevel.set_start(
                         level=level, timeout=STD_TIMEOUT
                     )
-                    self.assertEqual(ackcmd.identity, remote.salinfo.domain.identity)
+                    self.assertEqual(ackcmd.identity, remote.salinfo.identity)
                     print("Remote: wait for logLevel event")
                     data = await remote.evt_logLevel.next(
                         flush=False, timeout=STD_TIMEOUT